refactor: refactor product service
